+++ CHANGELOG.md
@@ -20,11 +20,13 @@ Change Category:
 + D-2: 在`model/unet.py`加入檢查model output的min/max值
 + B-9: 修改`model/utils.py`的儲存訓練快照格式
 + C-4: 新增`tfrecord2image_brats_2019_format.py`
-+ D-3: 補除以1.0，因tf.cast(data, tf.float32)沒有真的將data轉為float型態
++ ~D-3: 補除以1.0，因tf.cast(data, tf.float32)沒有真的將data轉為float型態~
++ D-3: 呼叫`model/unet.py`中的`self.output.eval()`出錯
 
 Future work:
 + B-10: 加入validation資料
 + C-5: 解構`model/read_tfrecord.py`，使可適應新增資料集
++ D-4: input mask有負值正常嗎？
 
 ### May 27, 2020
 + D-1: 發現transform_image.py並沒有將gif正確轉至jpg，label不是[0, 255]，而是漸變(不論使用pyplot, pillow, cv2儲存再讀取皆會出錯)，原因是因為jpg是有損壓縮，儲存後再載入會有差異，將label格式轉為png